api/pyproject.toml
[project]
name = "smb-dashboard-api"
version = "0.1.0"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "email-validator==2.2.0",
  "SQLAlchemy==2.0.35",
  "alembic==1.13.2",
  "psycopg[binary,pool]==3.2.1",
  "pandas==2.2.2",
  "numpy==1.26.4",
  "scikit-learn==1.5.2",
  "statsmodels==0.14.2",
  "redis==5.0.8",
  "rq==1.16.2",
  "boto3==1.34.156",
  "python-multipart==0.0.9",
  "orjson==3.10.7",
  "Mako==1.3.5"
]

[tool.uvicorn]
factory = false
